pyproject.toml
[tool.poetry]
package-mode = false
name = "streamlit-aggrid-examples"
version = "1.1.0"
description = ""
authors = ["Pablo Fonseca"]


[tool.poetry.dependencies]
python = ">=3.10"
streamlit-aggrid = ">=1.1.6"
streamlit-code-editor = "^0.1.12"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
